implementation of csv cont.

+++ solarpower/Solarpower.py
@@ -182,7 +182,7 @@ class Receiver (object):
             except IndexError:
                 Log("Wrong data format: {0[0]} {0[1]}".format(sys.exc_info()))
             else:
-                # csv.write(rrd_template, rrd_data)
+                my_csv.write(rrd_template, rrd_data)
                 try:
                     import rrdtool
                     rrdtool.update(RRDFILE, "--template", rrd_template, rrd_data)
@@ -199,22 +199,27 @@ class CSV (object):
 
     def new_file (self):
         self.filename = f"solarpower_{datetime.now().strftime('%Y%m%d')}.csv"
-        # if not file exists ... --> with open(...) ==> create new file
-        with open(self.filename, 'w', newline='') as file:
-            writer = csv.DictWriter(file, fieldnames=self.fieldnames)
-            writer.writeheader()
+        if not os.path.isfile(self.filename):
+            with open(self.filename, 'w', newline='') as file:
+                writer = csv.DictWriter(file, fieldnames=self.fieldnames)
+                writer.writeheader()
 
-    def get_item_from_rrd (self, item):
-        index = rrd_template.split(':').index(item)
-
+    def get_item_from_rrd (self, rrd_template, rrd_data, item):
+        return { item: rrd_data.split(':')[rrd_template.split(':').index(item)+1] }
 
     def write (self, rrd_template, rrd_data):
-        timestamp = datetime.now().strftime("%Y%m%d %H:%M:%S")
-        self.get_item_from_rrd(Main_Meter.field_P)
-        Log(f"Index of {Main_Meter.field_P}: {i}; rrd_template: {rrd_template}")
-        # with open(self.file, 'a', newline='') as file:
-        #     writer = csv.DictWriter(file, fieldnames=self.fieldnames)
-        #     writer.writerow({ "Timestamp": timestamp } | values)
+        """ python 3.9
+        csv_data = { "Timestamp": datetime.now().strftime("%Y%m%d %H:%M:%S") } |         \
+                   self.get_item_from_rrd(rrd_template, rrd_data, Main_Meter.field_P) |  \
+                   self.get_item_from_rrd(rrd_template, rrd_data, Solar_Meter.field_P)
+        """
+        csv_data = { "Timestamp": datetime.now().strftime("%Y%m%d %H:%M:%S") }
+        csv_data.update(self.get_item_from_rrd(rrd_template, rrd_data, Main_Meter.field_P))
+        csv_data.update(self.get_item_from_rrd(rrd_template, rrd_data, Solar_Meter.field_P))
+
+        with open(self.filename, 'a', newline='') as file:
+            writer = csv.DictWriter(file, fieldnames=self.fieldnames)
+            writer.writerow(csv_data)
 
 
 ###############################################################################
@@ -242,7 +247,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     if args.receiver:
-        csv = CSV()
+        my_csv = CSV()
         r = Receiver()
         r.start()
